+++ tests/repos/test_multi.py
@@ -1,13 +1,14 @@
 from unittest import mock
 
-import pkg_resources
 import pytest
+from packaging.requirements import Requirement
 
 from req_compile.containers import DistInfo
 from req_compile.errors import NoCandidateException
 from req_compile.repos import Repository
 from req_compile.repos.multi import MultiRepository
 from req_compile.repos.repository import Candidate
+from req_compile.utils import parse_version
 
 
 class FakeRepository(Repository):
@@ -63,7 +64,7 @@ def test_not_found():
     repo1 = FakeRepository("1")
     multi1 = MultiRepository(repo1)
     with pytest.raises(NoCandidateException):
-        multi1.get_dist(pkg_resources.Requirement("nonsense"))
+        multi1.get_dist(Requirement("nonsense"))
 
 
 def test_fetch_in_order():
@@ -74,20 +75,20 @@ def test_fetch_in_order():
 
     repo3.get_candidates.side_effect = lambda req: [
         Candidate(
-            "nonsense", ".", pkg_resources.parse_version("1.0"), None, None, "any", ""
+            "nonsense", ".", parse_version("1.0"), None, None, "any", ""
         )
     ]
     multi = MultiRepository(repo1, repo2, repo3)
 
-    result, cached = multi.get_dist(pkg_resources.Requirement("nonsense"))
+    result, cached = multi.get_dist(Requirement("nonsense"))
 
     assert repo1.get_candidates.called
     assert repo2.get_candidates.called
     assert repo3.get_candidates.called
 
-    assert result.version == pkg_resources.parse_version("1.0")
+    assert result.version == parse_version("1.0")
 
-    candidates = multi.get_candidates(pkg_resources.Requirement("nonsense"))
+    candidates = multi.get_candidates(Requirement("nonsense"))
     assert len(candidates) == 1
     assert candidates[0].name == "nonsense"
-    assert candidates[0].version == pkg_resources.parse_version("1.0")
+    assert candidates[0].version == parse_version("1.0")